Add burst support to normal ISCE2 workflow
